requirement.txt
aniso8601==8.0.0
asn1crypto==0.24.0
astroid==1.6.6
backports.functools-lru-cache==1.6.1
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==4.0.2
crcmod==1.7
cryptography==3.1.1
dill==0.2.9
Django==1.11.29
django-crispy-forms==1.9.0
djangorestframework==3.9.4
docutils==0.16
enum34==1.1.10
firebase==3.0.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.13.0
Flask-REST==1.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.16.0
futures==3.3.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-api-core==1.22.2
google-auth==1.22.0
google-cloud==0.34.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.2
google-gax==0.16.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.32.0
httplib2==0.18.1
idna==2.10
ipaddress==1.0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jws==0.1.3
keyring==10.6.0
keyrings.alt==3.0
Kivy-Garden==0.1.4
lazy-object-proxy==1.5.1
MarkupSafe==1.1.1
marshmallow==2.21.0
mccabe==0.6.1
mysql-connector-python==8.0.21
MySQL-python==1.2.5
mysqlclient==1.3.10
oauth2client==3.0.0
oauthlib==3.1.0
Pillow==6.2.2
ply==3.8
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.4.3
Pygments==2.5.2
pygobject==3.26.1
PyJWT==1.7.1
pylint==1.9.5
PyMySQL==0.10.1
Pyrebase==3.0.27
python-jwt==2.0.1
python-openid==2.2.5
python-resize-image==1.1.19
python-social-auth==0.3.6
pytz==2020.1
pyxdg==0.25
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.5
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
